pyproject.toml
[tool.black]
line-length = 79

[tool.poetry]
name = "allure-steplog-example"
version = "0.1.0"
description = ""
authors = ["Mikhail Ushanov <gm.mephisto@gmail.com>"]
packages = [
    { include = "allure_test" },
]
include = ["tests"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.4"
allure-pytest = "^2.9.43"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.5b2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
